Here's the 3D position/camera stuff
The old camera code I wrote about is now PositionRotation
It applies to camera and models. It can be controlled in the debug build if you want. Its main use is to apply lerp/slerp to move around smoothly
It's a class instead of a struct because of the annoying requirement to set every field before exiting a constructor
The animator class helps interpolate between two